Updated DataQuality Report and Created Data Review
Data Review documents what we know so far about each column as simply as possible. It could be used for reports to explain simply to the reader what each column means if they are new to the dataset.

It documents some issues with the data such as size and nulls. Its an ongoing document that should expand the more we learn about the data. We should add some case reviews to it such as a review about route 4